+++ src/main/java/com/endava/service_system/utils/ObjectArrayToServiceToUserDto.java
@@ -13,7 +13,7 @@ public class ObjectArrayToServiceToUserDto implements Converter<Object[],Service
     public ServiceToUserDto convert(Object[] source) {
         ServiceToUserDto serviceToUserDto=new ServiceToUserDto();
         serviceToUserDto.setCompanyName((String) source[0]);
-        serviceToUserDto.setId((Integer) source[1]);
+        serviceToUserDto.setId((Long) source[1]);
         serviceToUserDto.setTitle((String) source[2]);
         serviceToUserDto.setCategory((String) source[3]);
         serviceToUserDto.setDescription((String) source[4]);
